// lib/types.ts
export interface Contact {
  ID: string
  Group: string
  Check: boolean
  Conversation: string
  Name: string
  Phone_Number: string
  Email: string
  Definition1: string
  Definition2: string
  Definition3: string
  Memo: string
  Start_Or: string
  Whether_Or: string
  Checklist: string
  Dialogue_Name: string
  Change_Name: string
}

export interface Group {
  id: number
  name: string
}